# docs/swagger.yaml
components:
  schemas:
    models.ArmorResult:
      properties:
        armor:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
        set_name:
          type: string
        set_rarity:
          type: string
        stats:
          additionalProperties:
            type: number
          type: object
      type: object
    models.BestRunOutput:
      properties:
        damage_dealt:
          type: number
        damage_mitigated:
          type: number
        deaths:
          type: integer
        dungeon_class:
          type: string
        elapsed_time:
          type: integer
        grade:
          type: string
        mobs_killed:
          type: integer
        score_bonus:
          type: integer
        score_exploration:
          type: integer
        score_skill:
          type: integer
        score_speed:
          type: integer
        secrets_found:
          type: integer
        timestamp:
          type: integer
      type: object
...
    models.ClassData:
      properties:
        classAverage:
          type: number
        classAverageWithProgress:
          type: number
        classes:
          additionalProperties:
            $ref: '#/components/schemas/models.Skill'
          type: object
        selectedClass:
          type: string
        totalClassExp:
          type: number
      type: object
...
    models.DungeonFloorStats:
      properties:
        best_score:
          type: number
        fastest_time:
          type: number
        fastest_time_s:
          type: number
        fastest_time_s_plus:
          type: number
        milestone_completions:
          type: number
        mobs_killed:
          type: number
        most_damage:
          $ref: '#/components/schemas/models.MostDamageOutput'
        most_healing:
          type: number
        most_mobs_killed:
          type: number
        tier_completions:
          type: number
        times_played:
          type: number
        watcher_kills:
          type: number
      type: object
    models.DungeonStatsOutput:
      properties:
        bloodMobKills:
          type: integer
        highestFloorBeatenMaster:
          type: integer
        highestFloorBeatenNormal:
          type: integer
        secrets:
          $ref: '#/components/schemas/models.SecretsOutput'
      type: object
    models.DungeonsOutput:
      properties:
        catacombs:
          items:
            $ref: '#/components/schemas/models.FormattedDungeonFloor'
          type: array
          uniqueItems: false
        classes:
          $ref: '#/components/schemas/models.ClassData'
        level:
          $ref: '#/components/schemas/models.Skill'
        master_catacombs:
          items:
            $ref: '#/components/schemas/models.FormattedDungeonFloor'
          type: array
          uniqueItems: false
        stats:
          $ref: '#/components/schemas/models.DungeonStatsOutput'
      type: object
...
    models.EquipmentResult:
      properties:
        equipment:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
        stats:
          additionalProperties:
            type: number
          type: object
      type: object
...
    models.FormattedDungeonFloor:
      properties:
        best_run:
          $ref: '#/components/schemas/models.BestRunOutput'
        name:
          type: string
        stats:
          $ref: '#/components/schemas/models.DungeonFloorStats'
        texture:
          type: string
      type: object
...
    models.Gear:
      properties:
        armor:
          $ref: '#/components/schemas/models.ArmorResult'
        equipment:
          $ref: '#/components/schemas/models.EquipmentResult'
        wardrobe:
          items:
            items:
              $ref: '#/components/schemas/models.StrippedItem'
            type: array
          type: array
          uniqueItems: false
        weapons:
          $ref: '#/components/schemas/models.WeaponsResult'
      type: object
...
    models.MostDamageOutput:
      properties:
        damage:
          type: number
        type:
          type: string
      type: object
    models.OutputPets:
      properties:
        amount:
          type: integer
        amountSkins:
          type: integer
        missing:
          items:
            $ref: '#/components/schemas/models.StrippedPet'
          type: array
          uniqueItems: false
        petScore:
          $ref: '#/components/schemas/models.PetScore'
        pets:
          items:
            $ref: '#/components/schemas/models.StrippedPet'
          type: array
          uniqueItems: false
        total:
          type: integer
        totalCandyUsed:
          type: integer
        totalPetExp:
          type: integer
      type: object
...
    models.PetScore:
      properties:
        amount:
          type: integer
        reward:
          items:
            $ref: '#/components/schemas/models.PetScoreReward'
          type: array
          uniqueItems: false
        stats:
          additionalProperties:
            type: number
          type: object
      type: object
    models.PetScoreReward:
      properties:
        bonus:
          type: integer
        score:
          type: integer
        unlocked:
          type: boolean
      type: object
    models.PlayerResolve:
      properties:
        username:
          type: string
        uuid:
          type: string
      type: object
...
    models.ProcessingError:
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: string
      type: object
...
    models.SecretsOutput:
      properties:
        found:
          type: integer
        secretsPerRun:
          type: number
      type: object
    models.Skill:
      properties:
        level:
          type: integer
        levelCap:
          type: integer
        levelWithProgress:
          type: number
        maxLevel:
          type: integer
        maxed:
          type: boolean
        progress:
          type: number
        texture:
          type: string
        uncappedLevel:
          type: integer
        unlockableLevelWithProgress:
          type: number
        xp:
          type: integer
        xpCurrent:
          type: integer
        xpForNext:
          type: integer
      type: object
...
    models.StatsInfo:
      additionalProperties:
        type: integer
      type: object
...
    models.StrippedItem:
      properties:
        Count:
          type: integer
        containsItems:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
        display_name:
          type: string
        isInactive:
          type: boolean
        lore:
          items:
            type: string
          type: array
          uniqueItems: false
        rarity:
          type: string
        recombobulated:
          type: boolean
        shiny:
          type: boolean
        source:
          type: string
        texture_pack:
          type: string
        texture_path:
          type: string
        wiki:
          $ref: '#/components/schemas/models.WikipediaLinks'
      type: object
    models.StrippedPet:
      properties:
        active:
          type: boolean
        display_name:
          type: string
        level:
          type: integer
        lore:
          items:
            type: string
          type: array
          uniqueItems: false
        rarity:
          type: string
        stats:
          additionalProperties:
            type: number
          type: object
        texture_path:
          type: string
        type:
          type: string
      type: object
...
    models.WeaponsResult:
      properties:
        highest_priority_weapon:
          $ref: '#/components/schemas/models.StrippedItem'
        weapons:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
      type: object
    models.WikipediaLinks:
      properties:
        fandom:
          type: string
        official:
          type: string
      type: object
...
info:
  description: API for SkyCrypt - A Hypixel SkyBlock Stats Viewer
  title: SkyCrypt API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/dungeons/{uuid}/{profileId}:
    get:
      description: Returns dungeons for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.DungeonsOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
      summary: Get dungeons stats of a specified player
      tags:
      - dungeons
...
  /api/gear/{uuid}/{profileId}:
    get:
      description: Returns gear for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Gear'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get gear stats of a specified player
      tags:
      - gear
...
  /api/inventory/{uuid}/{profileId}/{inventoryId}:
    get:
      description: Returns inventory items for the given user, profile ID, and inventory
        ID. Supports museum, search, and other inventories.
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      - description: Inventory ID (e.g., museum, search, or other inventory types)
        in: path
        name: inventoryId
        required: true
        schema:
          type: string
      - description: Search string (required when inventoryId is 'search')
        in: path
        name: search
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.StrippedItem'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get inventory items for a specified player
      tags:
      - inventory
  /api/inventory/{uuid}/{profileId}/search/{search}:
    get:
      description: Returns inventory items for the given user, profile ID, and inventory
        ID. Supports museum, search, and other inventories.
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      - description: Inventory ID (e.g., museum, search, or other inventory types)
        in: path
        name: inventoryId
        required: true
        schema:
          type: string
      - description: Search string (required when inventoryId is 'search')
        in: path
        name: search
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.StrippedItem'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get inventory items for a specified player
      tags:
      - inventory
...
  /api/pets/{uuid}/{profileId}:
    get:
      description: Returns pets for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.OutputPets'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
      summary: Get pets stats of a specified player
      tags:
      - pets
  /api/playerStats/{uuid}/{profileId}:
    get:
      description: Returns player stats for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/models.StatsInfo'
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
      summary: Get player stats of a specified player
      tags:
      - playerStats
...
  /api/uuid/{username}:
    get:
      description: Returns the UUID associated with the given username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.PlayerResolve'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
      summary: Get UUID for a specified username
      tags:
      - uuid
servers:
- url: localhost:8080/